fix(GlobalExceptionHandler): Adds exception handlers to handle some error
handler
- AuthenticationException
- MethodArgumentNotValidException
- UsernameNotFoundException
